% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartogram_dorling.R
\name{cartogram_dorling_legend}
\alias{cartogram_dorling_legend}
\title{Dorling Cartogram Legend}
\usage{
cartogram_dorling_legend(
  x,
  var,
  pos = NULL,
  col = "white",
  border = "black",
  lwd = 1,
  values.cex = 0.6,
  values.round = 0,
  lty = 3,
  nb.circles = 4,
  title.txt = "Title of the legend",
  title.cex = 0.8,
  title.font = 2
)
}
\arguments{
\item{x}{dorling cartogram generetd with the function cartogram_dorling(). sf object.}

\item{var}{name of the vector containing the data}

\item{pos}{position of the legend. a vector of two coordinates in map units (c(x, y)}

\item{col}{fill color.}

\item{border}{stroke color.}

\item{lwd}{stroke thickness.}

\item{values.cex}{size of the values in the legend.}

\item{values.round}{rounding of numbers}

\item{lty}{line style}

\item{nb.circles}{number of circles in the legend}

\item{title.txt}{title of the legend.}

\item{title.cex}{size of the legend title.}

\item{title.font}{font of the legend title.}
}
\description{
Build and display the legend of a Dorling cartogram
}
\examples{
library(sf)
library(rnaturalearth)
library(cartogram)

# Data inport and Handling
world <- ne_countries(returnclass = "sf")
crs <- "+proj=eqc +lat_ts=0 +lat_0=0 +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
world <- st_transform(world, crs = crs)
world <- world[!is.na(world$pop_est),]

# Cartogram building
pop.dorling <- cartogram_dorling(world, "pop_est")
plot(st_geometry(pop.dorling), col ="red", border="white")

# Layout and Legend
title("World Population")
cartogram_dorling_legend(x = pop.dorling, pos <- NULL, var = "pop_est", col = "red",
                         border = "black", lwd = 1, values.cex = 0.6, values.round = 0,
                         lty = 3, nb.circles =  4, title.txt = "Number of inhabitants.",
                        title.cex = 0.8, title.font = 2)
}
\references{
Dorling, D. (1996). Area Cartograms: Their Use and Creation. In Concepts and Techniques in Modern Geography (CATMOG), 59.
}
