% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cartogram.R
\name{cartogramOutput}
\alias{cartogramOutput}
\title{Widget output function for use in Shiny}
\usage{
cartogramOutput(outputId, width = "100\%", height = "200px")
}
\description{
Widget output function for use in Shiny
}

