% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartogram_cont.R
\name{checkPolygonsGEOS}
\alias{checkPolygonsGEOS}
\title{Check polygons}
\usage{
checkPolygonsGEOS(obj, properly = TRUE, force = TRUE, useSTRtree = FALSE)
}
\arguments{
\item{obj}{Polygons object}

\item{properly}{use \code{\link[rgeos]{gContainsProperly}} rather than \code{\link[rgeos]{gContains}}}

\item{force}{ignore errors from \code{\link[rgeos]{createPolygonsComment}}}

\item{useSTRtree}{use \pkg{rgeos} STRtree in checking holes, which is much faster, but uses a lot of memory and does not release it on completion}
}
\description{
Code from \code{maptools 0.8-39}.
Copyright: Roger Bivand and Edzer Pebesma.
}
\keyword{internal}
