% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_cartogram.R
\name{d_cartogram}
\alias{d_cartogram}
\title{Calculate Dorling Cartogram Boundaries}
\usage{
d_cartogram(x, var, k = 5, weights = 1, itermax = 1000)
}
\arguments{
\item{x}{SpatialPolygonDataFrame or an sf object}

\item{var}{Name of the weighting variable in shp}

\item{k}{Share of the shp bounding box filled with the greatest circle}

\item{weights}{An optional vector of numeric weights (0 to 1 inclusive) to 
apply to the distance each circle moves during pair-repulsion. A weight of 0 
prevents any movement. A weight of 1 gives the default movement distance. A 
single value can be supplied for uniform weights. A vector with length less 
than the number of circles will be silently extended by repeating the final 
value. Any values outside the range [0, 1] will be clamped to 0 or 1.}

\item{itermax}{The maximum number of iterations.}
}
\value{
An object of the same class as shp with non overlaping proportional 
circles.
}
\description{
Construct a dorling cartogram.
}
\examples{
library(maptools)
library(cartogram)
library(rgdal)
data(wrld_simpl)

# Remove uninhabited regions
afr <- spTransform(wrld_simpl[wrld_simpl$REGION==2 & wrld_simpl$POP2005 > 0,], 
                   CRS("+init=epsg:3395"))

# Create cartogram
afr_carto <- d_cartogram(afr, "POP2005")

# Plot 
par(mfcol=c(1,2))
plot(afr, main="original")

plot(afr, main="distorted (sp)")
plot(afr_carto, col = "red", add=TRUE)

# Same with sf objects
library(sf)

afr_sf = st_as_sf(afr)

afr_sf_carto <- d_cartogram(afr_sf, "POP2005")

# Plot 
par(mfcol=c(1,3))
plot(afr, main="original")
plot(afr_carto, main="distorted (sp)")
plot(st_geometry(afr_sf_carto), main="distorted (sf)")
}
