% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cartogram.R
\name{renderCartogram}
\alias{renderCartogram}
\title{Widget render function for use in Shiny}
\usage{
renderCartogram(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

