% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cartogram.R
\name{cartogram}
\alias{cartogram}
\title{<Add Title>}
\usage{
cartogram(map = NULL, tooltip = NULL, circledata = NULL,
  scaledata = NULL, datasettransition = NULL,
  datasettransitioneasingfunction = NULL, zoom = NULL,
  mapprojection = NULL, colorbrewersetting = NULL, colordefault = NULL,
  mapscale = NULL, maptranslatex = NULL, maptranslatey = NULL,
  mapcenter = NULL, scalinginfo = NULL, message = NULL, width = NULL,
  height = NULL)
}
\description{
<Add Description>
}

