% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartogram_dorling.R
\name{cartogram_dorling}
\alias{cartogram_dorling}
\alias{cartogram_dorling.sf}
\alias{cartogram_dorling.SpatialPolygonsDataFrame}
\title{Calculate Non-Overlapping Circles Cartogram}
\usage{
cartogram_dorling(x, weight, k = 5, l = NULL, m_weight = 1, itermax = 1000)

\method{cartogram_dorling}{sf}(x, weight, k = 5, l = NULL, m_weight = 1, itermax = 1000)

\method{cartogram_dorling}{SpatialPolygonsDataFrame}(x, weight, k = 5, l = NULL, m_weight = 1, itermax = 1000)
}
\arguments{
\item{x}{SpatialPolygonsDataFrame, SpatialPointsDataFrame or an sf object}

\item{weight}{Name of the weighting variable in x}

\item{k}{Share of the bounding box of x filled by the larger circle}

\item{l}{Multiplying factor (in map units) to adjust the size of the circles.
Using l instead of k allows the cartograms to be made comparable with each other.}

\item{m_weight}{Circles' movements weights. An optional vector of numeric weights 
(0 to 1 inclusive) to 
apply to the distance each circle moves during pair-repulsion. A weight of 0 
prevents any movement. A weight of 1 gives the default movement distance. A 
single value can be supplied for uniform weights. A vector with length less 
than the number of circles will be silently extended by repeating the final 
value. Any values outside the range [0, 1] will be clamped to 0 or 1.}

\item{itermax}{Maximum iterations for the cartogram transformation.}
}
\value{
Non overlaping proportional circles of the same class as x.
}
\description{
Construct a cartogram which represents each geographic region 
as non-overlapping circles (Dorling 1996).
}
\examples{
library(sf)
library(rnaturalearth)
library(cartogram)

# Data inport and Handling
world <- ne_countries(returnclass = "sf")
crs <- "+proj=eqc +lat_ts=0 +lat_0=0 +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
world <- st_transform(world, crs = crs)
world <- world[!is.na(world$pop_est),]

# Cartogram building
pop.dorling <- cartogram_dorling(world, "pop_est")
plot(st_geometry(pop.dorling), col ="red", border="white")

# Layout and Legend
title("World Population")
}
\references{
Dorling, D. (1996). Area Cartograms: Their Use and Creation. In Concepts and Techniques in Modern Geography (CATMOG), 59.
}
